refactoring, creation of extra components, components rearrangement
