+++ code/src/components/styled/BackgroundSelection.js
